+++ data-raw/compounds-scorings.R
@@ -1,2 +1,2 @@
 compScorings <- read.csv(system.file("data-raw", "compounds-scorings.csv", package = "patRoon"), stringsAsFactors = FALSE)
-usethis::use_data(compScorings, internal = TRUE)
+usethis::use_data(compScorings, internal = TRUE, overwrite = TRUE)